requirements.txt
flask
anthropic
python-dotenv
Pillow
Flask-SQLAlchemy
psycopg2-binary
gunicorn
